requirements.txt
dotenv

html2text

langchain_community
langchain-openai

infinity_emb[all]

yfinance
pandas
pandas_datareader

beautifulsoup4
lxml

tqdm

Requests
httpx==0.23.3
httpx[http2]
httpx-cache[all]
tenacity

pypdf
PyMuPDF

numpy

setuptools
scipy

joblib

fake-useragent

playwright

fastapi
uvicorn